% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensysR.R
\name{checksysR}
\alias{checksysR}
\title{checksysR}
\arguments{
\item{Gamma0}{(input) The coefficient matrix \Gamma_0}

\item{Gamma1}{(input) The coefficient matrix \Gamma_1}

\item{Psi}{(input) The coefficient matrix \Psi}

\item{Pi}{(input) The coefficient matrix \Pi}

\item{C}{(input) The constant vector C}
}
\value{
Integer equal to the number of loose parameters. =0 when there is a unique solution, =-1 for no solution, >0 for indeterminacy
}
\description{
Use Sims (2002) method for checking for a unique solution a linear dynamic general equilibrium model
}
\details{
Model has the following form: \Gamma_0 x_t = C + \Gamma_1 x_t-1 + \Psi z_t + \Pi \eta_t

where...
 x_t is a vector of variables
 z_t is a vector of exogenous shocks
 \eta_t is a vector of endogenous expectation errors: \eta_t = x_t - E_t-1 x_t

 C is a vector of constants
 \Gamma_0 and \Gamma_1 are square matrices of coefficients, dimensions nvar x nvar
 \Psi is a matrix, dimension nvar x nshocks
 \Pi is a matrix, dimension nvar x nendo

The solution is given as,

x_t = D_sol + G_sol x_t-1 + M_sol z_t

The method includes a return value for existence and uniqueness of solution:
  Return value > 0: Indeterminacy, there are `return_value` number of loose endogenous variables, ignore outputs for D_{sol}, G_{sol}, and M_{sol}
  Return value = 0: Unique solution is found. There are 0 loose endogenous variables.
  Return value = -1: No solution exists.
}
